+++ sdk/resourcemanager/securityinsights/armsecurityinsights/zz_generated_response_types.go
@@ -324,6 +324,26 @@ type EntityRelationsClientGetRelationResponse struct {
 	Relation
 }
 
+// FileImportsClientCreateResponse contains the response from method FileImportsClient.Create.
+type FileImportsClientCreateResponse struct {
+	FileImport
+}
+
+// FileImportsClientDeleteResponse contains the response from method FileImportsClient.Delete.
+type FileImportsClientDeleteResponse struct {
+	FileImport
+}
+
+// FileImportsClientGetResponse contains the response from method FileImportsClient.Get.
+type FileImportsClientGetResponse struct {
+	FileImport
+}
+
+// FileImportsClientListResponse contains the response from method FileImportsClient.List.
+type FileImportsClientListResponse struct {
+	FileImportList
+}
+
 // IPGeodataClientGetResponse contains the response from method IPGeodataClient.Get.
 type IPGeodataClientGetResponse struct {
 	EnrichmentIPGeodata
@@ -498,6 +518,46 @@ func (p *ProductSettingsClientUpdateResponse) UnmarshalJSON(data []byte) error {
 	}
 	p.SettingsClassification = res
 	return nil
+}
+
+// SecurityMLAnalyticsSettingsClientCreateOrUpdateResponse contains the response from method SecurityMLAnalyticsSettingsClient.CreateOrUpdate.
+type SecurityMLAnalyticsSettingsClientCreateOrUpdateResponse struct {
+	SecurityMLAnalyticsSettingClassification
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type SecurityMLAnalyticsSettingsClientCreateOrUpdateResponse.
+func (s *SecurityMLAnalyticsSettingsClientCreateOrUpdateResponse) UnmarshalJSON(data []byte) error {
+	res, err := unmarshalSecurityMLAnalyticsSettingClassification(data)
+	if err != nil {
+		return err
+	}
+	s.SecurityMLAnalyticsSettingClassification = res
+	return nil
+}
+
+// SecurityMLAnalyticsSettingsClientDeleteResponse contains the response from method SecurityMLAnalyticsSettingsClient.Delete.
+type SecurityMLAnalyticsSettingsClientDeleteResponse struct {
+	// placeholder for future response values
+}
+
+// SecurityMLAnalyticsSettingsClientGetResponse contains the response from method SecurityMLAnalyticsSettingsClient.Get.
+type SecurityMLAnalyticsSettingsClientGetResponse struct {
+	SecurityMLAnalyticsSettingClassification
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type SecurityMLAnalyticsSettingsClientGetResponse.
+func (s *SecurityMLAnalyticsSettingsClientGetResponse) UnmarshalJSON(data []byte) error {
+	res, err := unmarshalSecurityMLAnalyticsSettingClassification(data)
+	if err != nil {
+		return err
+	}
+	s.SecurityMLAnalyticsSettingClassification = res
+	return nil
+}
+
+// SecurityMLAnalyticsSettingsClientListResponse contains the response from method SecurityMLAnalyticsSettingsClient.List.
+type SecurityMLAnalyticsSettingsClientListResponse struct {
+	SecurityMLAnalyticsSettingsList
 }
 
 // SentinelOnboardingStatesClientCreateResponse contains the response from method SentinelOnboardingStatesClient.Create.